Move transcript into separate file
